SMTPAPI Filters use "enable"
Toggling filters on and off is done with the `enable` key, not `enabled`.

If an `enable` key isn't present, the SMTPAPI assumes you're trying to enable it, and turns it on. This is why `enabled` will work when trying to turn on an app that is normally off, but will not work when trying to turn off an app that is normally on.