Cleaned up the tips at the top of the page
I am not sure about this change being the right move here. We have a problem where there are a couple caveats here that we need people to know about. In the original warning box, we had multiple instructions. It makes sense to delineate each item individually, but it does not make sense to have 4 info/warning boxes at the top of the page.