In depth explanation of nesting
I was originally confused by the statement "Section tags are an extension of substitution tags". I assumed that this meant "it works the same way" rather than "you must nest sections tags in substitution tags". These changes are added in order to attempt to clear up this confusion by changing the wording around and adding very specific substitution value tables to show what was replaced, how, and when by the API.