Change Global Suppression(s) to Global Unsubscribes
To reduce confusion I thought it would be best to use the same words and terminology that appears in the SendGrid dashboard. When the API docs talk about Global Suppressions there is a slight disconnect that this is referring to the Global Unsubscribes found in the Dashboard. In the User Guide Doc, under the Supporessions section, Global Unsubscribes are mentioned (which correlates to the Dahsboard). "Global Suppressions" isn't really found anywhere else in SendGrid's documention.