Unsubscribe link information clarification
I've added this extra phrase :
When creating a new campaign using own HTML code be sure to not delete the "To unsubscribe please click <a href="[unsubscribe]">here</a>" from the email body, otherwise SendGrid will NOT automatically add the unsubscribe link.

This is related to the automated addition of the unsubscribe link for the users who are using their own html code. Working previously in support I have noticed that there are very many clients that deleted the [unsubscribe] tab and then complaint that this is a SendGrid issue.