Add Accept header to docs
Some API calls require the Accept Header. if it's not provided, the error is not clear, it just shows "Authorization" type errors. This is EXTREMELY confusing to Customers & Support, causing a lot of wasted cycles while we try to figure out what's wrong with the API call ,when it turns out that it's just an undocumented required header.
I know there's a counter argument to fix the APi call, but until that's done, we need some way to know/remember what's wrong. the `Accept` header is never contra-indicated. it doesn't hurt to have it in there, so why no advise people to use it?