Update global.md
Current description is inaccurate and states that Global Stats show an aggregate for the parent and subusers. The Global Stats, in fact, only shows the data for the account currently being accessed (i.e. logged in as the Parent account, only stats for events passed through the parent account are considered and shown in the UI).