Update unsubscribe_groups.md - Opt In Language
Currently, the doc includes language that users are able to 'opt in' to various groups via the unsubscribe pages. This is confusing. Users are able to opt out of certain groups, however if a user does not opt out, this doesn't actually opt the user *in* to these communications. It just doesn't exclude them from them.

Language is updated to provide more clarity for the ability to manage only whether a user is explicitly opted out or not, rather than implying they are able to also opt in.