Update to Spam Check point
Added "Inbound Parse will check incoming emails for spam, then assign them a spam score and report. This will be reflected under `spam_score` and `spam_report` in your Parse Webhook post. The user can then determine how this score is interpreted and the severity."

Since we do not do anything based on spam score it, it is only reported and the user must take action based on what we post.