Update release_notes.md
Submitting the addition of a bug fix resolution, as well as the addition of search functionality to the custom field dropdown when creating a segment. 

PR also includes a fix to the icon set at the top of the page. The table class was applying incorrectly to the section, causing grey borders to appear in error. These are removed via inline CSS on the specific table element (as not to impact the proper table styling for entries below).