Update how-to-use-a-transactional-template-with-smtp.md
I have changed the word "transactional" to "legacy" as transactional templates means something different now. I have also created a warning, but not linked it to anything as of yet since the new dynamic templates have not been fully integrated into the documentation. Not sure if we have a way to change the URL from transactional templates to legacy templates.