Bounce Deletion Call Change
Added: {% info %} Please note, when you are completing the following requests they should not be combined. For example, if requesting to delete one address, it is not necessary to add the "delete_all=false" parameter in an effort to prevent deletion of further blocks, it is assumed. {% endinfo %} 

(Previously added to the Blocks API deletion documentation as well)