Remove Contradictory statement.
The statement "setting enabled status will apply settings from website" is both untrue, and directly contradicted by the following statement. If the filter is referenced _at all_ in the API string, the full state needs to be defined. As soon as we're looking to the API string for the settings, we don't look at the database. It's PUT-style, not PATCH.