update real-time to over time for segment updates
The real-time here through me off. I was expecting segments to update as soon as a contact was updated with the criteria needed to be segmented. Seems to be taking from 20 mins to over an hour for segments to update. 

It could be useful to add (within X number of hours) to give people an idea on how long it will take before their contacts start receiving emails from their automations tied to their segments, but I know that might be hard to predict.