Update references to Basic Auth username/password usage
As of 20 January 2021, Basic Authentication with account username + password combination will no longer work, documentation should be changed to require username: `apikey` and an API Key generated from the account as the password. If the integration does not permit a password of 69 characters (API Key length) that should be noted and further discussion should be had to determine SendGrid compatibility.