Fix user_name parameter typo in upgrading to API keys doc
The docs explaining the upgrade process to API keys for legacy Basic
Authentication incorrectly use a `username` parameter instead of
`user_name`. I was wondering why we were continuing to get lots of email
warnings from Sendgrid in the run-up to the deadline on Feb 17th. We
then started getting lots of ArgumentError exceptions with the message -
SMTP-AUTH requested but missing user name - Updating the parameter to
`user_name` fixed the issue.